% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{charvec_to_R}
\alias{charvec_to_R}
\title{Copy list of strings from Python to R character vector}
\usage{
charvec_to_R(name)
}
\description{
Copy list of strings from Python to R character vector
}
\examples{
pyrun("l = ['a', 'b']")
pyrun("print(l)")
charvec_to_R("l")
pyrun("l2 = [u'a', u'b']")
charvec_to_R("l2")
}

