% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{numvec_to_python}
\alias{numvec_to_python}
\title{Push data to python __main__ namespace}
\usage{
numvec_to_python(name, x)
}
\arguments{
\item{name}{Python variable name as string}

\item{x}{Numeric vector to copy to Python}
}
\description{
Push data to python __main__ namespace
}
\examples{
numvec_to_python("l", 1:10)
pyrun("print l")
}

