% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plot.R
\name{pyerrorbar}
\alias{pyerrorbar}
\title{Plot error bars using pyplot}
\usage{
pyerrorbar(x, y, xerr, yerr, args = NULL, show = FALSE)
}
\arguments{
\item{x}{A Numeric vector}

\item{y}{A Numeric vector}

\item{xerr}{A numeric vector of x errors at each data point}

\item{yerr}{A numeric vector of y errors at each data point}

\item{args}{String of extra arguments that are passed to Python}

\item{show}{If TRUE the plot will open and block (calls pyplot.show)}
}
\description{
Plot error bars using pyplot
}
\examples{
x = seq(0, 2*pi, length=50)
xerr = rep(0.05, length(x)) #absolute error
yerr = sin(x)*0.1 #relative error
pyerrorbar(x, sin(x), xerr, yerr, args=" fmt='o', ecolor='red' ")
if (interactive())
   pyshow()
}

