% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plot.R
\name{pyshow}
\alias{pyshow}
\title{Show figures
Run pyplot.show(). This will open all created figure windows and hang R before all windows are closed.}
\usage{
pyshow()
}
\description{
Show figures
Run pyplot.show(). This will open all created figure windows and hang R before all windows are closed.
}

