% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plot.R
\name{pystem}
\alias{pystem}
\title{Plot stem plot using pyplot}
\usage{
pystem(x, y = NULL, args = NULL, show = FALSE)
}
\arguments{
\item{x}{numeric vector containing the x coordinates of points}

\item{y}{numeric vector containing the y coordinates of points}

\item{args}{character string of further arguments passed to the **kwargs
argument of matplotlib.pyplot.scatter}

\item{show}{bool indicating whether to open a window with the plot}
}
\description{
Plot stem plot using pyplot
}
\examples{
x = seq(0, pi, length=20)
pystem(x, sin(x))
if(interactive())
   pyshow()
}

