% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/communication.R
\name{pyvar}
\alias{pyvar}
\title{Copy variable from R to Python}
\usage{
pyvar(name, x)
}
\arguments{
\item{name}{variable name in python}

\item{x}{R data, numeric and character vectors and numeric matrices are currently supported.}
}
\description{
Copy variable from R to Python
}
\examples{
pyvar("x", 1:10)
pyprint(x)
pyvar("s", c("Hello", "R!"))
pyprint(s)
pyvar(volcano) #Matrix
pyprint(volcano)
}

