% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plot.R
\name{pyvlines}
\alias{pyvlines}
\title{Plot vertical lines using pyplot.vlines}
\usage{
pyvlines(x, ymin, ymax, colors = "k", linestyles = "solid", label = "",
  args = NULL, show = FALSE)
}
\arguments{
\item{x}{numeric or a numeric vector with x-values for vertical lines}

\item{ymin}{numeric or a numeric vector of \code{length(x)} indicating lower
y-values of the lines}

\item{ymax}{numeric or a numeric vector of \code{length(x)} indicating upper
y-values of the lines}

\item{colors}{character indicating the color of the lines}

\item{label}{character string}

\item{args}{character string of further arguments passed to the **kwargs
argument of matplotlib.pyplot.vlines}

\item{show}{bool indicating whether to open a window with the plot}

\item{linestyle}{character string indicating the linestyle, one of
\code{"solid"} (default), \code{"dashed"}, \code{"dashdot"} or
\code{"dotted"}}
}
\description{
Plot vertical lines using pyplot.vlines
}
\examples{
pyvlines(c(1, 3, 5), 0, 10)
if (interactive()) pyshow()
}
\seealso{
\link{pyhlines}
  \url{http://matplotlib.org/api/pyplot_api.html#matplotlib.pyplot.vlines}
}

