\name{Rpyplot-package}
\alias{Rpyplot-package}
\alias{Rpyplot}
\docType{package}
\title{
R interaface to matplotlib
}

\description{Plot R vectors using matplotlib via Rcpp.}
\details{
\tabular{ll}{
Package: \tab Rpyplot\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2015-03-30\cr
License: \tab GPL (>= 2)\cr
}

}
\author{
Matti Pastell

Maintainer: Matti Pastell <matti.pastell@helsinki.fi>
}
\references{http://matplotlib.org}

\keyword{ package }
