% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/communication.R
\name{Rvar}
\alias{Rvar}
\title{Copy a variable from Python to R}
\usage{
Rvar(var)
}
\arguments{
\item{var}{Python variable name}
}
\description{
Copy a variable from Python to R
}
\examples{
pyvar("x", 1:10)
Rvar("x")
pyrun("f = 3")
Rvar("f")
pyrun("s = ['a', 'b', 'c']")
Rvar("s")
pyrun("s2 = 'a'")
Rvar("s2")
data(whiteside, package="MASS")
pyvar(whiteside)
Rvar(whiteside)
}

