% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plot.R
\name{pycontourf}
\alias{pycontourf}
\title{Filled contour using pyplot}
\usage{
pycontourf(x, y, z, args = NULL, show = FALSE)
}
\arguments{
\item{x}{A Numeric vector}

\item{y}{A Numeric vector}

\item{z}{array of z-axis coordinates}

\item{args}{String of extra arguments that are passed to Python}

\item{show}{If TRUE the plot will open and block (calls pyplot.show)}
}
\description{
Filled contour using pyplot
}
\examples{
x <- 1:nrow(volcano)*.2
y <- 1:ncol(volcano)*.5
pycontourf(x, y, volcano)
if (interactive())
   pyshow()
}

