% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plot.R
\name{pyerrorbar}
\alias{pyerrorbar}
\title{Plot error bars using pyplot}
\usage{
pyerrorbar(x, y, xerr, yerr, args = NULL, show = FALSE)
}
\arguments{
\item{x}{numeric vector containing the x coordinates of points}

\item{y}{numeric vector containing the y coordinates of points}

\item{xerr}{A numeric vector of x errors at each data point}

\item{yerr}{A numeric vector of y errors at each data point}

\item{args}{character string of further arguments passed to the **kwargs
argument of matplotlib.pyplot.scatter}

\item{show}{bool indicating whether to open a window with the plot}
}
\description{
Plot error bars using pyplot
}
\examples{
x = seq(0, 2*pi, length=50)
xerr = rep(0.05, length(x)) #absolute error
yerr = sin(x)*0.1 #relative error
pyerrorbar(x, sin(x), xerr, yerr, args=" fmt='o', ecolor='red' ")
if (interactive())
   pyshow()
}

