% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plot.R
\name{pyhlines}
\alias{pyhlines}
\title{Plot horizontal lines using pyplot.hlines}
\usage{
pyhlines(y, xmin, xmax, colors = "k", linestyles = "solid", label = "",
  args = NULL, show = FALSE)
}
\arguments{
\item{y}{numeric or a numeric vector with y-values for horizontal lines}

\item{xmin}{numeric or a numeric vector of \code{length(y)} indicating lower
x-values of the lines}

\item{xmax}{numeric or a numeric vector of \code{length(y)} indicating upper
x-values of the lines}

\item{colors}{character indicating the color of the lines}

\item{label}{character string}

\item{args}{character string of further arguments passed to the **kwargs
argument of matplotlib.pyplot.hlines}

\item{show}{bool indicating whether to open a window with the plot}

\item{linestyle}{character string indicating the linestyle, one of
\code{"solid"} (default), \code{"dashed"}, \code{"dashdot"} or
\code{"dotted"}}
}
\description{
Plot horizontal lines using pyplot.hlines
}
\examples{
pyhlines(1:10, 0, 10)
if (interactive()) pyshow()
}
\seealso{
\link{pyvlines}
  \url{http://matplotlib.org/api/pyplot_api.html#matplotlib.pyplot.hlines}
}

