% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plot.R
\name{pyplot}
\alias{pyplot}
\title{Plot x vs y}
\usage{
pyplot(x, y = NULL, args = NULL, show = FALSE)
}
\arguments{
\item{x}{A Numeric vector}

\item{y}{A Numeric vector}

\item{args}{String of extra arguments that are passed to Python}

\item{show}{If TRUE the plot will open and block (calls pyplot.show)}
}
\description{
Plot 2D line plots and scatter plots using pyplot.plot. You can interact with the plots using matplotlib tools
after calling pyshow() (=pyplot.show)
}
\examples{
x = seq(0, 2*pi, length=100)
pyplot(x, sin(x))
pyplot(x, cos(x), args="'r--', linewidth=3")
pyfigure()
pyplot(x, sin(x)^2)
xlabel("x")
ylabel("$sin(x)^2$")
pyshow()
}

