% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plot.R
\name{pystem}
\alias{pystem}
\title{Plot stem plot using pyplot}
\usage{
pystem(x, y = NULL, args = NULL, show = FALSE)
}
\arguments{
\item{x}{A Numeric vector}

\item{y}{A Numeric vector}

\item{args}{String of extra arguments that are passed to Python}

\item{show}{If TRUE the plot will open and block (calls pyplot.show)}
}
\description{
Plot stem plot using pyplot
}
\examples{
x = seq(0, pi, length=20)
pystem(x, sin(x))
pyshow()
}

