% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plot.R
\name{pyvlines}
\alias{pyvlines}
\title{Plot vertical lines using pyplot.vlines}
\usage{
pyvlines(x, ymin, ymax, colors = "k", linestyles = "solid", label = "",
  args = NULL, show = FALSE)
}
\arguments{
\item{x}{numeric or a numeric vector with x-values for vertical lines}

\item{ymin}{numeric or a numeric vector of \code{length(x)} indicating lower
y-values of the lines}

\item{ymax}{numeric or a numeric vector of \code{length(x)} indicating upper
y-values of the lines}

\item{colors}{color of the line(s) (default: \code{"k"}), can be one of
\itemize{
  \item{single character for basic built-in matplotlib colors, see
    \url{http://matplotlib.org/api/colors_api.html#module-matplotlib.colors
    }
  }
  \item{a numeric value between 0 and 1 as character, indicating gray
    shade, e.g. \code{"0.75"}}
  \item{hex string, e.g. \code{"#00ff00"} for green}
  \item{character string with HTML color name, e.g. \code{"slateblue"},
    see \url{http://www.w3schools.com/html/html_colornames.asp}}
  \item{a character vector of \code{length(x)} containing a color
    specification as described above for each point separately}
  \item{a numeric vector of \code{length(x)}, where each numeric is used
    to map the according point to a colormap using the \code{cmap},
    \code{norm}, \code{vmin} and \code{vmax} arguments}
  \item{a numeric matrix with \code{length(x)} rows and three columns. Each
    row represents one point in \code{x}/\code{y}, the rows set values for
    RGB (red, green and blue, each between 0 and 1)}
  \item{a numeric matrix with \code{length(x)} rows and four columns. Each
    row represents one point in \code{x}/\code{y}, the rows set values for
    RGBA (red, green, blue and alpha, each between 0 and 1)}
}}

\item{label}{character string}

\item{args}{character string of further arguments passed to the **kwargs
argument of matplotlib.pyplot.vlines}

\item{show}{bool indicating whether to open a window with the plot}

\item{linestyle}{character string or character vector of \code{length(x)}
indicating the linestyle(s), using \code{"solid"} (default),
\code{"dashed"}, \code{"dashdot"} or \code{"dotted"}}
}
\description{
Plot vertical lines using pyplot.vlines
}
\examples{
pyvlines(c(1, 3, 5), 0, 10, colors = c("r", "g", "b"),
         linestyles = c("solid", "dashed", "dotted"))
if (interactive()) pyshow()
}
\seealso{
\link{pyhlines}
  \url{http://matplotlib.org/api/pyplot_api.html#matplotlib.pyplot.vlines}
}

